Move tests into unit test on fcda-binding-list
